util 기능 모듈화
